Created controller to calculate ranking

+++ asm/src/main/kotlin/com/github/marciovmartins/futsitev3/ranking/domain/PointCriteria.kt
@@ -1,7 +1,7 @@
 package com.github.marciovmartins.futsitev3.ranking.domain
 
 data class PointCriteria(
-    val victories: Int,
-    val draws: Int,
-    val defeats: Int,
+    val victories: Long,
+    val draws: Long,
+    val defeats: Long,
 )
