+++ asm/src/main/kotlin/com/github/marciovmartins/futsitev3/ranking/usecase/PointCriteriaDTO.kt
@@ -1,7 +1,7 @@
 package com.github.marciovmartins.futsitev3.ranking.usecase
 
 data class PointCriteriaDTO(
-    val victories: Int,
-    val draws: Int,
-    val defeats: Int
+    val victories: Long,
+    val draws: Long,
+    val defeats: Long
 )
